SystemUI: Avoid "phantom" status bar collapsing
There is one specific case where the status bar may collapse
for no apparent reason: When the last notification is removed.

If Quick Settings panel is showing while such a notification
is removed, the status bar WILL collapse -- and the user may
be left wondering why, because the notification that was just
removed isn't visible in that case.

A way to reproduce this behaviour follows:
  1) Dismiss all notifications, even the sticky ones.
  2) Open the Quick Settings Panel and tap the appropriate
     tile to light the Torch
  3) Tap the tile again, to extinguish the Torch.

As you can see, the Quick Settings Panel collapses. This
happens because Torch puts up a notification, when lit and
removes it, just after it's turned off.

This will happen to whatever notification that hides itself,
e.g., WiFi AP status when WiFi AP is turned off. 
This too will trigger this issue. (Of course, only in case
there are no other notifications left)

-----

This patch prevents the bar from collapsing, as a consequence
of removing the last notification, when Quick Settings is
showing.

Original fix by: ThiagoVinicius
https://github.com/Root-Box/frameworks_base/commit/b4440e3630e9e51cdc5c29fa1e2936fd3eb0174b