HW Key Actions for Recent Tasks key (hardware and virtual) 1) Hardware key: some devices have hardware key mapped to APP_SWITCH key code (HTC One) which is handled in PhoneWindowManager's interceptKeyBeforeDispatching and launches recent apps dialog. We override its behaviour when either single-tap or long-press is defined in GB settings
2) Virtual key: this one is tricky since virtual recent apps key has no
key code assigned and thus is not processed within PhoneWindowManager.
We dynamically assign  APP_SWITCH key code to it as soon as either
single-tap or long-press action is defined. This makes virtual key fire
key events that get processed within PhoneWindowManager overriding
default behaviour.
When both, single-tap and long-press are set to default, key code is
reset to 0 which reverts key to default behaviour